# app/server/generated/routers/workflow.py
# generated by fastapi-codegen:
#   filename:  api.yaml
#   timestamp: 2024-05-09T02:31:37+00:00

from __future__ import annotations

from fastapi import APIRouter, FastAPI

from ..dependencies import *
from ..handlers import workflow_handler

router = APIRouter(tags=['workflow'])


@router.get(
    '/review/{jurisdictionId}/{caseId}',
    response_model=ReviewMeta,
    responses={'424': {'model': Error}},
    tags=['workflow'],
)
def get_review_documents(
    jurisdiction_id: str = Path(..., alias='jurisdictionId'),
    case_id: str = Path(..., alias='caseId'),
) -> Union[ReviewMeta, Error]:
    """
    Get a list of documents to review
    """
    return workflow_handler.get_review_documents(**locals())